% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SEB}
\alias{SEB}
\title{Partitioning Data into Statistically Equivalent Blocks}
\usage{
SEB(X, nints, intervals = TRUE, order = "successive")
}
\arguments{
\item{X}{a numeric matrix.}

\item{nints}{a vector of positive integers specifying into how many intervals each variable (column of \code{X}) should be cut.}

\item{intervals}{logical. Whether not only observations should be assigned to blocks, but also a full characterization of the blocks should be returned. Default is \code{TRUE}.}

\item{order}{a character string specifying how to proceed in cutting axes, must be either "\code{successive}" or "\code{random}". Default is "\code{successive}".}
}
\value{
A list of length one or two containing the following components.

\item{labels}{An integer vector of block labels for each observation (row of \code{X}).}
\item{intervals}{Returned if \code{intervals} is \code{TRUE}. A numeric matrix with \code{prod(nints)} rows and \code{2 * ncol(X)} columns describing each block. For any variable \code{k} from 1 to \code{ncol(X)}, columns \code{2 * k - 1} and \code{2 * k} contain intervals, of the form (a, b], to which variable \code{k} belongs to in the corresponding block.}
}
\description{
Partitioning Data into Statistically Equivalent Blocks
}
\examples{
n <- 20
k <- 2
# Generating a matrix of n observations and k variables
X <- matrix(runif(n * k), ncol = k)

# Defining the number of intervals to which each variable should be partitioned
nints <- c(2, 3)

# Partitioning X into prod(nints) statistically equivalent blocks
SEB(X, nints)

# Choosing at random whether first to cut the first or the second variable.
SEB(X, nints, order = "random")

# Returning only the block labels
SEB(X, nints, FALSE)
}
\references{
M.P. Gessaman \emph{A Consistent Nonparametric Multivariate Density Estimator Based on Statistically Equivalent Blocks} The Annals of Mathematical Statistics (1970) 1344--1346
}
\author{
Julius Vainora
}
